GCW-3700 added update-user-image component
