+++ app/templates/manage_account.hbs
@@ -1,4 +1,4 @@
-<nav class="tab-bar user-details-heading">
+<nav class="tab-bar user-details-heading accounts-page">
   <section class="left-small">
     {{#link-to "app_menu_list" class="user-back-link"}}
     <i class="back_icon" aria-hidden="true">{{fa-icon 'angle-left'}}</i>
@@ -11,41 +11,13 @@
   </section>
 </nav>
 
-<section class="main-section add-user-form new-user-section ui">
+<section class="main-section add-user-form new-user-section ui accounts-form">
     <div class="row">
       <div class="small-5 columns item-label">
         {{t 'users.image'}}
       </div>
       <div class="small-7 columns">
-        <span class={{if user.image "image" "image-box" }}>
-          {{#if user.image}}
-          <img src={{user.image.faceCaptureImageUrl}} class="thumb" />
-          {{else}}
-          <a {{action "triggerUpload" }} class="{{unless isReady 'disabled'}}">
-            <i aria-hidden="true">{{fa-icon 'camera'}}</i><br />
-            {{t 'users.add_image'}}
-          </a>
-          {{cloudinary-upload ready="uploadReady" progress="uploadProgress"
-          always="uploadComplete" done="uploadEditedImageSuccess" submit="uploadStart"}}
-          {{/if}}
-        </span>
-
-        <div class="item-label image-edit">
-          {{#if user.image}}
-          <div>
-            <a {{action "triggerUpload" }} class="{{unless isReady 'disabled'}}">
-              {{t 'users.edit_image'}}
-            </a>
-            {{cloudinary-upload ready="uploadReady" progress="uploadProgress"
-            always="uploadComplete" done="uploadEditedImageSuccess" submit="uploadStart"}}
-          </div>
-          <div class="delete-image">
-            <a {{action "deleteImage" }}>
-              {{t 'users.delete_image'}}
-            </a>
-          </div>
-          {{/if}}
-        </div>
+        {{upload-user-image user=user}}
       </div>
     </div>
